Switch to double duration
This was done since the float precision wasn't enough for some systems

+++ include/transition.hpp
@@ -40,7 +40,7 @@ namespace squi {
 		float progress{1};
 
 	public:
-		TransitionData(float *value, std::chrono::duration<float> &duration, Curve curve);
+		TransitionData(float *value, std::chrono::duration<double> &duration, Curve curve);
 
 		void to(const float &newValue);
